+++ packages/lumx-react/src/components/text-field/RawInputText.tsx
@@ -1,31 +1,26 @@
-import { ComponentProps, ChangeEventHandler, SyntheticEvent, useRef, useCallback } from 'react';
+import { SyntheticEvent } from 'react';
 
 import { Theme, useTheme } from '@lumx/react';
 
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
-import { useMergeRefs } from '@lumx/react/utils/react/mergeRefs';
-import { HasClassName, HasTheme } from '@lumx/react/utils/type';
-import { classNames } from '@lumx/core/js/utils';
+import {
+    RawInputText as UI,
+    RawInputTextProps as UIProps,
+    DEFAULT_PROPS,
+} from '@lumx/core/js/components/TextField/RawInputText';
 
-import { INPUT_NATIVE_CLASSNAME } from './constants';
-
-const { block } = classNames.bem(INPUT_NATIVE_CLASSNAME);
-
-type NativeInputProps = Omit<ComponentProps<'input'>, 'value' | 'onChange'>;
+import { ReactToJSX } from '@lumx/react/utils/type/ReactToJSX';
 
 /**
  * Defines the props of the component.
  */
-export interface RawInputTextProps extends NativeInputProps, HasTheme, HasClassName {
-    value?: string;
+export interface RawInputTextProps extends ReactToJSX<UIProps> {
     onChange?: (value: string, name?: string, event?: SyntheticEvent) => void;
 }
 /**
  * Component default props.
  */
-export const DEFAULT_PROPS: Partial<RawInputTextProps> = {
-    type: 'text',
-};
+export { DEFAULT_PROPS };
 
 /**
  * Raw input text component
@@ -33,39 +28,12 @@ export const DEFAULT_PROPS: Partial<RawInputTextProps> = {
  */
 export const RawInputText = forwardRef<RawInputTextProps, HTMLInputElement>((props, ref) => {
     const defaultTheme = useTheme() || Theme.light;
-    const {
-        className,
-        theme = defaultTheme,
-        value,
-        onChange,
-        type = DEFAULT_PROPS.type,
-        name,
-        ...forwardedProps
-    } = props;
-    const textareaRef = useRef<HTMLInputElement>(null);
+    const { theme = defaultTheme, onChange, ...restOfProps } = props;
 
-    const handleChange: ChangeEventHandler<HTMLInputElement> = useCallback(
-        (evt) => {
-            onChange?.(evt.target.value, name, evt);
-        },
-        [onChange, name],
-    );
-
-    return (
-        <input
-            {...forwardedProps}
-            name={name}
-            type={type}
-            ref={useMergeRefs(ref, textareaRef)}
-            className={classNames.join(
-                className,
-                block({
-                    [`theme-${theme}`]: Boolean(theme),
-                    text: true,
-                }),
-            )}
-            onChange={handleChange}
-            value={value}
-        />
-    );
+    return UI({
+        ...restOfProps,
+        ref,
+        theme,
+        handleChange: onChange,
+    });
 });